issue-1444: create CpuWaitFailure counter in filestore/blockstore actors
