+++ cloud/blockstore/libs/storage/volume_balancer/volume_balancer_actor.h
@@ -40,6 +40,7 @@ private:
     NMonitoring::TDynamicCounters::TCounterPtr InitiallyPreempted;
 
     NMonitoring::TDynamicCounters::TCounterPtr CpuWait;
+    NMonitoring::TDynamicCounters::TCounterPtr CpuWaitFailure;
 
     std::unique_ptr<TVolumeBalancerState> State;
