+++ cloud/storage/core/libs/diagnostics/cgroup_stats_fetcher.cpp
@@ -1,10 +1,6 @@
 #include "cgroup_stats_fetcher.h"
 
-#include <cloud/storage/core/libs/diagnostics/critical_events.h>
 #include <cloud/storage/core/libs/diagnostics/logging.h>
-#include <cloud/storage/core/libs/diagnostics/monitoring.h>
-
-#include <library/cpp/monlib/dynamic_counters/counters.h>
 
 #include <util/datetime/cputimer.h>
 #include <util/generic/yexception.h>
@@ -14,8 +10,6 @@
 #include <util/system/file.h>
 
 namespace NCloud::NStorage {
-
-using namespace NMonitoring;
 
 namespace {
 
@@ -28,9 +22,7 @@ private:
     const TString ComponentName;
 
     const ILoggingServicePtr Logging;
-    const IMonitoringServicePtr Monitoring;
     const TString StatsFile;
-    const TCgroupStatsFetcherMonitoringSettings MonitoringSettings;
 
     TLog Log;
 
@@ -38,20 +30,14 @@ private:
 
     TDuration Last;
 
-    TIntrusivePtr<NMonitoring::TCounterForPtr> FailCounter;
-
 public:
     TCgroupStatsFetcher(
             TString componentName,
             ILoggingServicePtr logging,
-            IMonitoringServicePtr monitoring,
-            TString statsFile,
-            TCgroupStatsFetcherMonitoringSettings monitoringSettings)
+            TString statsFile)
         : ComponentName(std::move(componentName))
         , Logging(std::move(logging))
-        , Monitoring(std::move(monitoring))
         , StatsFile(std::move(statsFile))
-        , MonitoringSettings(std::move(monitoringSettings))
     {
     }
 
@@ -64,13 +50,11 @@ public:
                 StatsFile,
                 EOpenModeFlag::OpenExisting | EOpenModeFlag::RdOnly);
         } catch (...) {
-            ReportCpuWaitFatalError();
             STORAGE_ERROR(BuildErrorMessageFromException());
             return;
         }
 
         if (!CpuAcctWait.IsOpen()) {
-            ReportCpuWaitFatalError();
             STORAGE_ERROR("Failed to open " << StatsFile);
             return;
         }
@@ -98,7 +82,6 @@ public:
             constexpr i64 bufSize = 1024;
 
             if (CpuAcctWait.GetLength() >= bufSize - 1) {
-                ReportCpuWaitFatalError();
                 CpuAcctWait.Close();
                 return MakeError(E_INVALID_STATE, StatsFile + " is too large");
             }
@@ -124,7 +107,6 @@ public:
 
             return retval;
         } catch (...) {
-            ReportCpuWaitFatalError();
             auto errorMessage = BuildErrorMessageFromException();
             CpuAcctWait.Close();
             return MakeError(E_FAIL, std::move(errorMessage));
@@ -136,24 +118,6 @@ public:
         auto msg = TStringBuilder() << "IO error for " << StatsFile;
         msg << " with exception " << CurrentExceptionMessage();
         return msg;
-    }
-
-    void ReportCpuWaitFatalError()
-    {
-        if (FailCounter) {
-            return;
-        }
-        if (MonitoringSettings.ComponentGroupName.empty() ||
-            MonitoringSettings.CountersGroupName.empty() ||
-            MonitoringSettings.CounterName.empty())
-        {
-            return;
-        }
-        FailCounter = Monitoring->GetCounters()
-            ->GetSubgroup("counters", MonitoringSettings.CountersGroupName)
-            ->GetSubgroup("component", MonitoringSettings.ComponentGroupName)
-            ->GetCounter(MonitoringSettings.CounterName, false);
-        *FailCounter = 1;
     }
 };
 
@@ -183,16 +147,12 @@ struct TCgroupStatsFetcherStub final
 ICgroupStatsFetcherPtr CreateCgroupStatsFetcher(
     TString componentName,
     ILoggingServicePtr logging,
-    IMonitoringServicePtr monitoring,
-    TString statsFile,
-    TCgroupStatsFetcherMonitoringSettings settings)
+    TString statsFile)
 {
     return std::make_shared<TCgroupStatsFetcher>(
         std::move(componentName),
         std::move(logging),
-        std::move(monitoring),
-        std::move(statsFile),
-        std::move(settings));
+        std::move(statsFile));
 }
 
 ICgroupStatsFetcherPtr CreateCgroupStatsFetcherStub()
@@ -209,5 +169,33 @@ TString BuildCpuWaitStatsFilename(const TString& serviceName)
     }
     return {};
 }
+
+// One can use either a service name and have the stats file inferred from it,
+// or provide the stats file explicitly.
+NCloud::NStorage::ICgroupStatsFetcherPtr BuildCgroupStatsFetcher(
+    const TString& cpuWaitServiceName,
+    const TString& cpuWaitFilename,
+    const TLog& log,
+    ILoggingServicePtr logging,
+    TString componentName)
+{
+    if (cpuWaitServiceName.empty() && cpuWaitFilename.empty()) {
+        const auto& Log = log;
+        STORAGE_INFO(
+            "CpuWaitServiceName and CpuWaitFilename are empty, can't build "
+            "CgroupStatsFetcher");
+        return CreateCgroupStatsFetcherStub();
+    }
+
+    TString statsFile =
+        cpuWaitFilename.empty()
+            ? NCloud::NStorage::BuildCpuWaitStatsFilename(cpuWaitServiceName)
+            : cpuWaitFilename;
+
+    return CreateCgroupStatsFetcher(
+        std::move(componentName),
+        std::move(logging),
+        statsFile);
+};
 
 }   // namespace NCloud::NStorage
