+++ cloud/storage/core/libs/diagnostics/cgroup_stats_fetcher_ut.cpp
@@ -1,8 +1,5 @@
 #include "cgroup_stats_fetcher.h"
 
-#include "critical_events.h"
-
-#include <cloud/storage/core/libs/diagnostics/monitoring.h>
 #include <cloud/storage/core/libs/diagnostics/logging.h>
 
 #include <library/cpp/monlib/dynamic_counters/counters.h>
@@ -17,17 +14,6 @@ namespace {
 ////////////////////////////////////////////////////////////////////////////////
 
 const TString ComponentName = "STORAGE_CGROUPS";
-
-auto SetupCriticalEvents(IMonitoringServicePtr monitoring)
-{
-    auto rootGroup = monitoring->GetCounters()
-        ->GetSubgroup("counters", "storage");
-
-    auto serverGroup = rootGroup->GetSubgroup("component", "server");
-    InitCriticalEventsCounter(serverGroup);
-
-    return serverGroup;
-}
 
 void UpdateCGroupWaitDuration(TTempFileHandle& file, TDuration value)
 {
@@ -48,8 +34,6 @@ Y_UNIT_TEST_SUITE(TCGroupStatFetcherTest)
 {
     Y_UNIT_TEST(ShouldReadStats)
     {
-        auto monitoring = CreateMonitoringServiceStub();
-        auto serverGroup = SetupCriticalEvents(monitoring);
         TTempFileHandle statsFile("test");
 
         UpdateCGroupWaitDuration(statsFile, TDuration::MicroSeconds(10));
@@ -57,13 +41,7 @@ Y_UNIT_TEST_SUITE(TCGroupStatFetcherTest)
         auto fetcher = CreateCgroupStatsFetcher(
             ComponentName,
             CreateLoggingService("console"),
-            monitoring,
-            statsFile.Name(),
-            {
-                .CountersGroupName = "storage",
-                .ComponentGroupName = "test",
-                .CounterName = "CpuWaitFailure",
-            });
+            statsFile.Name());
         fetcher->Start();
 
         auto cpuWait = fetcher->GetCpuWait();
@@ -80,37 +58,15 @@ Y_UNIT_TEST_SUITE(TCGroupStatFetcherTest)
             cpuWait.GetResult());
 
         fetcher->Stop();
-
-        UNIT_ASSERT_VALUES_EQUAL(
-            0,
-            serverGroup
-                ->GetCounter("AppCriticalEvents/CpuWaitCounterReadError", true)
-                ->Val());
     }
 
     Y_UNIT_TEST(ShouldReportErrorIfFileIsMissing)
     {
-        auto monitoring = CreateMonitoringServiceStub();
-        auto serverGroup = SetupCriticalEvents(monitoring);
-
         auto fetcher = CreateCgroupStatsFetcher(
             ComponentName,
             CreateLoggingService("console"),
-            monitoring,
-            "noname",
-            {
-                .CountersGroupName = "storage",
-                .ComponentGroupName = "server",
-                .CounterName = "CpuWaitFailure",
-            });
+            "noname");
         fetcher->Start();
-
-        auto failCounter = monitoring->GetCounters()
-            ->GetSubgroup("counters", "storage")
-            ->GetSubgroup("component", "server")
-            ->GetCounter("CpuWaitFailure", false);
-
-        UNIT_ASSERT_VALUES_EQUAL(1, failCounter->Val());
 
         auto cpuWait = fetcher->GetCpuWait();
         UNIT_ASSERT_C(HasError(cpuWait), cpuWait.GetError());
@@ -122,32 +78,18 @@ Y_UNIT_TEST_SUITE(TCGroupStatFetcherTest)
 
     Y_UNIT_TEST(ShouldReportErrorIfNewValueIsLowerThanPrevious)
     {
-        auto monitoring = CreateMonitoringServiceStub();
-        auto serverGroup = SetupCriticalEvents(monitoring);
         TTempFileHandle statsFile("test");
-
         UpdateCGroupWaitDuration(statsFile, TDuration::MicroSeconds(100));
 
         auto fetcher = CreateCgroupStatsFetcher(
             ComponentName,
             CreateLoggingService("console"),
-            monitoring,
-            "test",
-            {
-                .CountersGroupName = "storage",
-                .ComponentGroupName = "test",
-                .CounterName = "CpuWaitFailure",
-            });
+            "test");
         fetcher->Start();
 
         UpdateCGroupWaitDuration(statsFile, TDuration::MicroSeconds(80));
         auto cpuWait = fetcher->GetCpuWait();
         UNIT_ASSERT_C(HasError(cpuWait), cpuWait.GetError());
-        UNIT_ASSERT_VALUES_EQUAL(
-            1,
-            serverGroup
-                ->GetCounter("AppCriticalEvents/CpuWaitCounterReadError", true)
-                ->Val());
 
         UpdateCGroupWaitDuration(statsFile, TDuration::MicroSeconds(100));
         cpuWait = fetcher->GetCpuWait();
@@ -155,11 +97,6 @@ Y_UNIT_TEST_SUITE(TCGroupStatFetcherTest)
         UNIT_ASSERT_VALUES_EQUAL(
             TDuration::MicroSeconds(20),
             cpuWait.GetResult());
-        UNIT_ASSERT_VALUES_EQUAL(
-            1,
-            serverGroup
-                ->GetCounter("AppCriticalEvents/CpuWaitCounterReadError", true)
-                ->Val());
 
         fetcher->Stop();
     }
